pyproject.toml
[project]
name = "ai-init"
version = "0.1.0"
description = "First steps to AI using OPEN AI api, langchain and langgraph"
authors = [
    {name = "EdYar"}
]
readme = "README.md"
requires-python = ">=3.11, <4.0.0"
dependencies = [
    "langchain==1.0.2",
    "langgraph==1.0.1",
    "openai==2.6.1",
    "python-dotenv (>=1.2.1,<2.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false